// src/app/models/new-property.ts
export interface NewProperty {
  title: string;
  image_path: string;
  description: string;
  n_rooms: number;
  n_bathrooms: number;
  sqm: number;
  price: number;
  priority: number;
  street: string;
  owner_id: string;
  house_number: string;
  neighborhood: string;
  zip_code: string;
  city_name: string;
  state_id: number;
  latitude: string;
  longitude: string;
  property_type_id: number;
  disposition_type_id: number;
}